% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_tree.R
\name{add_wgd_rects}
\alias{add_wgd_rects}
\title{Add rectangles indicating WGD events in a phylogenetic tree}
\usage{
add_wgd_rects(p, tree, wgd_dates, rh = 0.25, highlight = NULL)
}
\arguments{
\item{p}{A `ggtree/ggplot` object containing a phylogenetic tree 
visualization.}

\item{tree}{A `phylo` object with the species tree used to create
the plot in \strong{p}.}

\item{wgd_dates}{Data frame with WGD date information.}

\item{rh}{Numeric indicating the height of the rectangles representing
WGD events. Default: 0.15.}

\item{highlight}{A character vector with IDs of WGD events to highlight.}
}
\value{
A `ggtree` object.
}
\description{
Add rectangles indicating WGD events in a phylogenetic tree
}
\examples{
data(tree)
data(species_metadata)
data(wgd_dates)
keep <- species_metadata[species_metadata$family == "Fabaceae", "latin_name"]

ftree <- ape::keep.tip(tree, keep)

p <- plot_timetree_rectangular(ftree, species_metadata)
add_wgd_rects(p, ftree, wgd_dates, rh = 0.3)

}
