% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_clades}
\alias{available_clades}
\title{Available clades that users can select from}
\format{
A character vector with all possible clades associated with species
in this app.
}
\usage{
data(available_clades)
}
\description{
Available clades that users can select from
}
\examples{
data(available_clades)
}
\keyword{datasets}
