% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_urls}
\alias{data_urls}
\title{Data frame with URLs to original data (CDS and annotation) for each species}
\format{
A data frame with the following variables:
\describe{
  \item{species_name}{Character, latin name for each species.}
  \item{family}{Character, family for each species.}
  \item{order}{Character, order for each species.}
  \item{clade}{Character, one of 'ANA grade', 'Monocots', 'Magnoliids',
  'Chloranthales', 'Ceratophyllales', or 'Eudicots'.}
  \item{CDS_URL}{Character, URL to FASTA file with CDS.}
  \item{Annotation_URL}{Character, URL to GFF3 file with annotation.}
}
}
\usage{
data(data_urls)
}
\description{
Data frame with URLs to original data (CDS and annotation) for each species
}
\examples{
data(data_urls)
}
\keyword{datasets}
