% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nodepath_and_age}
\alias{get_nodepath_and_age}
\title{Get the node path from a tip of the tree to another node}
\usage{
get_nodepath_and_age(tree, tip, node = 467)
}
\arguments{
\item{tree}{A `phylo` object with an ultrametric species tree.}

\item{tip}{Character indicating the label of tr he tip to start from.}

\item{node}{Numeric indicating the ID of the node where the path ends.}
}
\value{
A 3-column data frame with the following variables:
\itemize{
  \strong{from} numeric, node ID.
  \strong{to} numeric, node ID.
  \strong{age} numeric, node age (distance from the tip).
}
}
\description{
Get the node path from a tip of the tree to another node
}
\examples{
data(tree)
df <- get_nodepath_and_age(tree, "Glycine_max", 467)
}
