% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{periods}
\alias{periods}
\title{Period data from the International Commission on Stratigraphy (v2023/06)}
\format{
A data frame with 22 rows and 5 variables. See `?deeptime::periods`
for more details.
}
\usage{
data(periods)
}
\description{
Same data set as in \strong{deeptime::periods}. See `?deeptime::periods`
for more details.
}
\examples{
data(periods)
}
\keyword{datasets}
