% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_distros.R
\name{plot_consensus_age_distro}
\alias{plot_consensus_age_distro}
\title{Plot a histogram and density line with a consensus posterior distributions of WGD ages}
\usage{
plot_consensus_age_distro(hist_data)
}
\arguments{
\item{hist_data}{A data frame of pre-computed histogram data for 
a specific WGD as in package data \strong{hist_data}.}
}
\value{
A ggplot object.
}
\description{
Plot a histogram and density line with a consensus posterior distributions of WGD ages
}
\examples{
data(posterior_hist)

hist_data <- posterior_hist$combined |> dplyr::filter(WGD_ID == "MUSA_beta")
p <- plot_consensus_age_distro(hist_data)
}
