% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_metadata}
\alias{species_metadata}
\title{Taxonomic information for all 466 species in \strong{tree}}
\format{
A data frame with the following variables:
\describe{
  \item{latin_name}{Character, latin name for each species.}
  \item{family}{Character, family for each species.}
  \item{order}{Character, order for each species.}
  \item{taxonomy_1}{Character, one of 'Commenilids', 'Lamiids',
  'Campanulids', 'Fabids', or 'Malvids'.}
  \item{taxonomy_2}{Character, one of 'Asterids' or 'Rosids'.}
  \item{taxonomy_3}{Character, one of 'ANA grade', 'Monocots', 'Magnoliids',
  'Chloranthales', 'Ceratophyllales', or 'Eudicots'.}
}
}
\usage{
data(species_metadata)
}
\description{
Taxonomic information for all 466 species in \strong{tree}
}
\examples{
data(species_metadata)
}
\keyword{datasets}
