% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_tree}
\alias{subset_tree}
\title{Subset tree to include only a specific clade and outgroup}
\usage{
subset_tree(tree, metadata, clade)
}
\arguments{
\item{tree}{A `phylo` object with a species tree.}

\item{metadata}{A data frame of species metadata with taxonomic information.}

\item{clade}{Character indicating which clade to subset.}
}
\value{
A `phylo` object with a subset of the original tree.
}
\description{
Subset tree to include only a specific clade and outgroup
}
\examples{
data(tree)
data(species_metadata)

final_tree <- subset_tree(tree, species_metadata, "Fabaceae")
}
