% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_wgd_per_clade}
\alias{subset_wgd_per_clade}
\title{Subset a data frame of WGD events to include only species in a clade}
\usage{
subset_wgd_per_clade(wgd_table, metadata, clade)
}
\arguments{
\item{wgd_table}{A data frame with WGD events and statistics as 
in package data \strong{wgd_dates}.}

\item{metadata}{A data frame of species metadata with taxonomic information.}

\item{clade}{Character indicating which clade to subset.}
}
\value{
A data frame with WGD events for species in selected clade.
}
\description{
Subset a data frame of WGD events to include only species in a clade
}
\examples{
data(wgd_dates)
data(species_metadata)

subset_wgd_per_clade(wgd_dates, species_metadata, clade = "Fabales")
}
