% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{Time-calibrated species tree for 466 angiosperm species}
\format{
An object of class `phylo` with 466 tips and 465 internal nodes
}
\usage{
data(tree)
}
\description{
Time-calibrated species tree for 466 angiosperm species
}
\examples{
data(tree)
}
\keyword{datasets}
