% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgd_dates}
\alias{wgd_dates}
\title{WGD events and ages for all 466 species in \strong{tree}}
\format{
A data frame with the following variables:
\describe{
  \item{wgd_id}{Character, unique ID of WGD events.}
  \item{phylogenetic_location}{Character, branch of the plant phylogeny 
  where WGD event is.}
  \item{species}{Character, species name.}
  \item{ks_peak}{Numeric, Ks peak.}
  \item{credible_range}{Character, credible range for the Ks peak.}
  \item{posterior_mean}{Numeric, mean of the posterior distribution of
  WGD ages.}
  \item{posterior_median}{Numeric, median of the posterior distribution
  of WGD ages.}
  \item{posterior_mode}{Numeric, mode of the posterior distribution of
  WGD ages.}
  \item{x90_percent_hpd}{Character, range with 90\% HPD.}
  \item{consensus_peak}{Numeric, peak of the consensus distribution
  of WGD ages.}
  \item{consensus_mean}{Numeric, mean of the consensus distribution of
  WGD ages.}
  \item{x90_percent_hcr}{Character, range with 90\% HCR.}
  \item{full_species}{Character with the names of all species sharing that 
  WGD in their history.}
}
}
\usage{
data(wgd_dates)
}
\description{
WGD events and ages for all 466 species in \strong{tree}
}
\examples{
data(wgd_dates)
}
\keyword{datasets}
